+++ sdk/core/core-rest-pipeline/src/index.ts
@@ -20,7 +20,7 @@ export {
   RawHttpHeadersInput,
   RequestBodyType,
   SendRequest,
-  StreamBuilder,
+  StreamProducer,
   TlsSettings,
   TransferProgressEvent,
 } from "./interfaces";