+++ sdk/openai/openai/src/OpenAIClient.ts
@@ -43,7 +43,7 @@ import {
 } from "./models/audio.js";
 import { nonAzurePolicy } from "./api/policies/nonAzure.js";
 import { getAudioTranscription, getAudioTranslation } from "./api/operations.js";
-import { StreamBuilder } from "@azure/core-rest-pipeline";
+import { StreamProducer } from "@azure/core-rest-pipeline";
 
 export { OpenAIClientOptions } from "./api/OpenAIContext.js";
 
@@ -223,7 +223,7 @@ export class OpenAIClient {
    */
   async getAudioTranscription(
     deploymentName: string,
-    buildFileStream: StreamBuilder,
+    buildFileStream: StreamProducer,
     options?: GetAudioTranscriptionOptions
   ): Promise<AudioResultSimpleJson>;
   /**
@@ -236,14 +236,14 @@ export class OpenAIClient {
    */
   async getAudioTranscription<Format extends AudioResultFormat>(
     deploymentName: string,
-    buildFileStream: StreamBuilder,
+    buildFileStream: StreamProducer,
     format: Format,
     options?: GetAudioTranscriptionOptions
   ): Promise<AudioResult<Format>>;
   // implementation
   async getAudioTranscription<Format extends AudioResultFormat>(
     deploymentName: string,
-    buildFileStream: StreamBuilder,
+    buildFileStream: StreamProducer,
     formatOrOptions?: Format | GetAudioTranscriptionOptions,
     inputOptions?: GetAudioTranscriptionOptions
   ): Promise<AudioResult<Format>> {
@@ -277,7 +277,7 @@ export class OpenAIClient {
    */
   async getAudioTranslation(
     deploymentName: string,
-    buildFileStream: StreamBuilder,
+    buildFileStream: StreamProducer,
     options?: GetAudioTranslationOptions
   ): Promise<AudioResultSimpleJson>;
   /**
@@ -290,14 +290,14 @@ export class OpenAIClient {
    */
   async getAudioTranslation<Format extends AudioResultFormat>(
     deploymentName: string,
-    buildFileStream: StreamBuilder,
+    buildFileStream: StreamProducer,
     format: Format,
     options?: GetAudioTranslationOptions
   ): Promise<AudioResult<Format>>;
   // implementation
   async getAudioTranslation<Format extends AudioResultFormat>(
     deploymentName: string,
-    buildFileStream: StreamBuilder,
+    buildFileStream: StreamProducer,
     formatOrOptions?: Format | GetAudioTranslationOptions,
     inputOptions?: GetAudioTranslationOptions
   ): Promise<AudioResult<Format>> {