+++ sdk/openai/openai/src/api/operations.ts
@@ -55,7 +55,7 @@ import {
   GetAudioTranslationOptions,
 } from "../models/audio.js";
 import { renameKeysToCamelCase } from "./util.js";
-import { StreamBuilder, createFileFromStream } from "@azure/core-rest-pipeline";
+import { StreamProducer, createFileFromStream } from "@azure/core-rest-pipeline";
 
 export function _getEmbeddingsSend(
   context: Client,
@@ -672,7 +672,7 @@ function _getChatCompletionsSendX(
 export async function getAudioTranslation(
   context: Client,
   deploymentName: string,
-  buildFileStream: StreamBuilder,
+  buildFileStream: StreamProducer,
   options?: GetAudioTranslationOptions
 ): Promise<AudioResultSimpleJson>;
 /**
@@ -687,7 +687,7 @@ export async function getAudioTranslation(
 export async function getAudioTranslation<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  buildFileStream: StreamBuilder,
+  buildFileStream: StreamProducer,
   format: Format,
   options?: GetAudioTranslationOptions
 ): Promise<AudioResult<Format>>;
@@ -695,7 +695,7 @@ export async function getAudioTranslation<Format extends AudioResultFormat>(
 export async function getAudioTranslation<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  buildFileStream: StreamBuilder,
+  buildFileStream: StreamProducer,
   formatOrOptions?: Format | GetAudioTranslationOptions,
   inputOptions?: GetAudioTranslationOptions
 ): Promise<AudioResult<Format>> {
@@ -735,7 +735,7 @@ export async function getAudioTranslation<Format extends AudioResultFormat>(
 export async function getAudioTranscription(
   context: Client,
   deploymentName: string,
-  buildFileStream: StreamBuilder,
+  buildFileStream: StreamProducer,
   options?: GetAudioTranscriptionOptions
 ): Promise<AudioResultSimpleJson>;
 /**
@@ -750,7 +750,7 @@ export async function getAudioTranscription(
 export async function getAudioTranscription<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  buildFileStream: StreamBuilder,
+  buildFileStream: StreamProducer,
   format: Format,
   options?: GetAudioTranscriptionOptions
 ): Promise<AudioResult<Format>>;
@@ -758,7 +758,7 @@ export async function getAudioTranscription<Format extends AudioResultFormat>(
 export async function getAudioTranscription<Format extends AudioResultFormat>(
   context: Client,
   deploymentName: string,
-  buildFileStream: StreamBuilder,
+  buildFileStream: StreamProducer,
   formatOrOptions?: Format | GetAudioTranscriptionOptions,
   inputOptions?: GetAudioTranscriptionOptions
 ): Promise<AudioResult<Format>> {